+++ CHANGELOG.md
@@ -43,7 +43,7 @@ Our versioning strategy is as follows:
 
 ### 🐛 Bug Fixes
 
-* [sitecore-jss-react] Use ComponentDataOverride instead of ComponentProps for BYOCComponent ([#1682](https://github.com/Sitecore/jss/pull/1682))
+* [sitecore-jss-react] Use ComponentDataOverride instead of ComponentProps for BYOCComponent ([#1682](https://github.com/Sitecore/jss/pull/1682))([#1688](https://github.com/Sitecore/jss/pull/1688))
 * `[sitecore-jss-proxy]` Setting "followRedirects" to "true" breaks HEAD requests ([#1630](https://github.com/Sitecore/jss/pull/1630))
 * `[templates/nextjs-sxa]` Fix shown horizontal scrollbar in EE mode. ([#1625](https://github.com/Sitecore/jss/pull/1625)), ([#1626](https://github.com/Sitecore/jss/pull/1626))
 * `[sitecore-jss-nextjs]` Fix issue when redirects works each every other times. ([#1629](https://github.com/Sitecore/jss/pull/1629))